Extract PostedBy component for reusable seller links
- Create independent PostedBy component in frontend/src/components/
- Refactor SearchablesProfile to use new PostedBy component
- Add PostedBy component to DownloadableSearchableDetails
- Maintain existing functionality while improving code reusability
